// apps/booking/src/order/enum/order-vehicle-status.enum.ts
export enum OrderVehicleStatus {
  PENDING_OWNER_RESPONSE = 'PENDING_OWNER_RESPONSE',
  OWNER_ACCEPTED_DRIVER_REASSIGNMENT = 'OWNER_ACCEPTED_DRIVER_REASSIGNMENT',
  DRIVER_PENDING = 'DRIVER_PENDING',
  DRIVER_ACCEPTED = 'DRIVER_ACCEPTED',
  DRIVER_REJECTED = 'DRIVER_REJECTED',
  OWNER_REJECTED = 'OWNER_REJECTED',
  OWNER_CANCELLED = 'OWNER_CANCELLED',
  ARRIVED = 'ARRIVED',
  IN_TRANSIT = 'IN_TRANSIT',
  COMPLETED = 'COMPLETED',
}